+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - Changed async operation interface types to never be `typing.Optional` ([issue#11]).
+- Interface type hints are now `typing.Protocol`.
 
 [issue#11]: https://github.com/pywinrt/pywinrt/issues/11